# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Education Planner API",
    "description": "\n    ## Динамический планировщик образовательной траектории\n    \n    API для студентов УрФУ, помогающее:\n    - 📚 Выбирать специальность и курсы\n    - 📊 Отслеживать развитие навыков\n    - 🎯 Получать персональные рекомендации\n    - 🔄 Моделировать смену специальности\n    \n    ### Основные возможности:\n    \n    - **Аутентификация**: JWT токены (access + refresh)\n    - **Специальности**: Просмотр roadmap и матрицы навыков\n    - **Курсы**: Запись, прохождение, отслеживание прогресса\n    - **Навыки**: Древовидная структура, gap-анализ\n    - **Рекомендации**: Умные подсказки по выбору курсов\n    - **What-If анализ**: Моделирование смены специальности\n    \n    ### Для начала работы:\n    1. Зарегистрируйтесь через `/api/auth/register`\n    2. Получите токен через `/api/auth/login`\n    3. Используйте токен в заголовке: `Authorization: Bearer \u003Ctoken\u003E`\n    ",
    "contact": {
      "name": "Education Planner Team",
      "email": "support@edu-planner.ru"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/courses/": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get All Courses",
        "description": "Получить список курсов с фильтрами",
        "operationId": "get_all_courses_api_courses__get",
        "parameters": [
          {
            "name": "specialization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по специальности",
              "title": "Specialization Id"
            },
            "description": "Фильтр по специальности"
          },
          {
            "name": "semester",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по семестру",
              "title": "Semester"
            },
            "description": "Фильтр по семестру"
          },
          {
            "name": "is_elective",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по типу курса",
              "title": "Is Elective"
            },
            "description": "Фильтр по типу курса"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseResponse"
                  },
                  "title": "Response Get All Courses Api Courses  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recommendations/courses": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Course Recommendations",
        "description": "Получить рекомендации курсов.\nНа первом этапе возвращает случайные курсы, восполняющие дефицит навыков.",
        "operationId": "get_course_recommendations_api_recommendations_courses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "description": "Количество рекомендаций",
              "default": 5,
              "title": "Limit"
            },
            "description": "Количество рекомендаций"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recommendations/what-if": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "What If Specialization Change",
        "description": "Моделирование смены специальности.\nПоказывает, как изменится профиль навыков при переходе на другую специальность.",
        "operationId": "what_if_specialization_change_api_recommendations_what_if_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatIfResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/minors/": {
      "get": {
        "tags": [
          "minors"
        ],
        "summary": "Get All Minors",
        "description": "Получить список всех майнеров.\n\n- **skip**: количество записей для пропуска (пагинация)\n- **limit**: максимальное количество записей\n- **include_inactive**: включать ли неактивные майнеры",
        "operationId": "get_all_minors_api_minors__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Minor"
                  },
                  "title": "Response Get All Minors Api Minors  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/minors/{minor_id}": {
      "get": {
        "tags": [
          "minors"
        ],
        "summary": "Get Minor By Id",
        "description": "Получить детальную информацию о майнере.\n\nВключает:\n- целевые навыки майнера\n- курсы, входящие в майнер\n- агрегированные теги",
        "operationId": "get_minor_by_id_api_minors__minor_id__get",
        "parameters": [
          {
            "name": "minor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Minor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinorDetailed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/minors/recommendations/for-me": {
      "get": {
        "tags": [
          "minors"
        ],
        "summary": "Get Minor Recommendations",
        "description": "Получить рекомендации майнеров для текущего студента.\n\nРекомендации основаны на:\n- дифференцированных тегах (с весами)\n- пройденных курсах студента\n- интересах и профиле студента\n\nВозвращает майнеры, отсортированные по score (релевантности).",
        "operationId": "get_minor_recommendations_api_minors_recommendations_for_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "description": "Количество рекомендаций",
              "default": 5,
              "title": "Limit"
            },
            "description": "Количество рекомендаций"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinorRecommendationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/minors/recommendations/similar": {
      "get": {
        "tags": [
          "minors"
        ],
        "summary": "Get Similar Minors",
        "description": "Получить рекомендации похожих майнеров при смене текущего.\n\nИспользуется, когда студент хочет сменить майнер —\nсистема предлагает похожие направления на основе:\n- текущего выбранного майнера\n- уже пройденных курсов",
        "operationId": "get_similar_minors_api_minors_recommendations_similar_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "current_minor_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ID текущего майнера",
              "title": "Current Minor Id"
            },
            "description": "ID текущего майнера"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "default": 3,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinorRecommendationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/minors/select": {
      "post": {
        "tags": [
          "minors"
        ],
        "summary": "Select Minor",
        "description": "Выбрать майнер для студента.\n\nЕсли у студента уже выбран другой майнер (status=selected),\nстарый майнер будет переведён в статус 'archived',\nа все полученные навыки останутся в резюме.",
        "operationId": "select_minor_api_minors_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserMinorCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMinor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/minors/{minor_id}/complete": {
      "post": {
        "tags": [
          "minors"
        ],
        "summary": "Complete Minor",
        "description": "Отметить майнер как завершённый.\n\nПосле завершения майнера:\n- навыки остаются в профиле студента\n- майнер отмечается как 'completed' в резюме\n- можно выбрать новый майнер",
        "operationId": "complete_minor_api_minors__minor_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "minor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Minor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMinor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/minors/my/history": {
      "get": {
        "tags": [
          "minors"
        ],
        "summary": "Get My Minors",
        "description": "Получить историю майнеров студента.\n\nВключает:\n- текущий выбранный (selected)\n- завершённые (completed)\n- архивные (archived)",
        "operationId": "get_my_minors_api_minors_my_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserMinor"
                  },
                  "type": "array",
                  "title": "Response Get My Minors Api Minors My History Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "credits": {
            "type": "integer",
            "maximum": 10,
            "minimum": 1,
            "title": "Credits",
            "default": 3
          },
          "semester": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "is_elective": {
            "type": "boolean",
            "title": "Is Elective",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "specialization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          },
          "difficulty_levels": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Levels"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "CourseResponse",
        "description": "Схема ответа с курсом"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Minor": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "minor_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minor Type"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "is_active"
        ],
        "title": "Minor",
        "description": "Полная схема майнера"
      },
      "MinorDetailed": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "minor_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minor Type"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "target_skills": {
            "items": {
              "$ref": "#/components/schemas/MinorSkillDetailed"
            },
            "type": "array",
            "title": "Target Skills",
            "default": []
          },
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseWithTags"
            },
            "type": "array",
            "title": "Courses",
            "default": []
          },
          "aggregated_tags": {
            "type": "object",
            "title": "Aggregated Tags",
            "description": "Теги с суммарными весами"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "is_active"
        ],
        "title": "MinorDetailed",
        "description": "Детальная информация о майнере с навыками и курсами"
      },
...
      "MinorRecommendationsResponse": {
        "properties": {
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/MinorRecommendation"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "recommendations",
          "total_count"
        ],
        "title": "MinorRecommendationsResponse",
        "description": "Ответ с рекомендациями майнеров"
      },
...
      "RecommendationsResponse": {
        "properties": {
          "specialization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          },
          "specialization_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Name"
          },
          "recommended_courses": {
            "items": {
              "$ref": "#/components/schemas/RecommendedCourse"
            },
            "type": "array",
            "title": "Recommended Courses"
          }
        },
        "type": "object",
        "required": [
          "recommended_courses"
        ],
        "title": "RecommendationsResponse",
        "description": "Ответ с рекомендациями курсов"
      },
      "RecommendedCourse": {
        "properties": {
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Skills"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "course_name",
          "reason",
          "priority",
          "missing_skills"
        ],
        "title": "RecommendedCourse",
        "description": "Рекомендованный курс"
      },
...
      "UserMinor": {
        "properties": {
          "minor_id": {
            "type": "integer",
            "title": "Minor Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "minor_id",
          "id",
          "user_id",
          "status"
        ],
        "title": "UserMinor",
        "description": "Полная схема выбора майнера"
      },
      "UserMinorCreate": {
        "properties": {
          "minor_id": {
            "type": "integer",
            "title": "Minor Id"
          }
        },
        "type": "object",
        "required": [
          "minor_id"
        ],
        "title": "UserMinorCreate",
        "description": "Схема для выбора майнера"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WhatIfRequest": {
        "properties": {
          "new_specialization_id": {
            "type": "integer",
            "title": "New Specialization Id"
          }
        },
        "type": "object",
        "required": [
          "new_specialization_id"
        ],
        "title": "WhatIfRequest",
        "description": "Запрос на моделирование смены специальности"
      },
      "WhatIfResponse": {
        "properties": {
          "current_specialization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Specialization Id"
          },
          "current_specialization_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Specialization Name"
          },
          "new_specialization_id": {
            "type": "integer",
            "title": "New Specialization Id"
          },
          "new_specialization_name": {
            "type": "string",
            "title": "New Specialization Name"
          },
          "current_completion_percentage": {
            "type": "number",
            "title": "Current Completion Percentage"
          },
          "new_completion_percentage": {
            "type": "number",
            "title": "New Completion Percentage"
          },
          "skills_comparison": {
            "items": {
              "$ref": "#/components/schemas/WhatIfSkillComparison"
            },
            "type": "array",
            "title": "Skills Comparison"
          },
          "courses_to_keep": {
            "type": "integer",
            "title": "Courses To Keep"
          },
          "total_completed_courses": {
            "type": "integer",
            "title": "Total Completed Courses"
          }
        },
        "type": "object",
        "required": [
          "current_specialization_id",
          "current_specialization_name",
          "new_specialization_id",
          "new_specialization_name",
          "current_completion_percentage",
          "new_completion_percentage",
          "skills_comparison",
          "courses_to_keep",
          "total_completed_courses"
        ],
        "title": "WhatIfResponse",
        "description": "Результат моделирования смены специальности"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}